+++ jsscanner/analysis/__init__.py
@@ -4,5 +4,6 @@ from .static import StaticAnalyzer
 from .secrets import SecretScanner
 from .unpacking import BundleUnpacker
 from .filtering import NoiseFilter
+from .semgrep import SemgrepAnalyzer
 
-__all__ = ['StaticAnalyzer', 'SecretScanner', 'BundleUnpacker', 'NoiseFilter']
+__all__ = ['StaticAnalyzer', 'SecretScanner', 'BundleUnpacker', 'NoiseFilter', 'SemgrepAnalyzer']